experimental branch. Multiple arity fns always emit IFn protocol properties. Single method non-variadic fns now also emit IFn protocol property as well via emit-single-fn-method.
In emit :invoke we now
have a special case for IFn protocol, we don't pass this as the
first arg, this is to avoid any overhead from extra dispatching just
to drop the this argument.

Note this means IFn will need special handling in the deftype/record
case as well - we'll need to drop the extra argument when we get
there. Currently it's not an issue since IFn is handled specially by
emitting a call / apply methods to the prototype - not -invoke
properties.

Related changes to the deftype/record emitting macros.